+++ bcolz/chunked_eval.py
@@ -13,7 +13,6 @@ from __future__ import absolute_import
 
 import sys
 import math
-import inspect
 
 import numpy as np
 import bcolz
@@ -35,12 +34,8 @@ def is_sequence_like(var):
     return False
 
 
-def _getvars(expression, user_dict, depth, vm):
-    """Get the variables in `expression`.
-
-    `depth` specifies the depth of the frame in order to reach local
-    or global variables.
-    """
+def _getvars(expression, user_dict, vm):
+    """Get the variables in `expression`."""
 
     cexpr = compile(expression, '<string>', 'eval')
     if vm in ("python", "dask"):
@@ -56,15 +51,14 @@ def _getvars(expression, user_dict, depth, vm):
 
     # Get the local and global variable mappings of the user frame
     user_locals, user_globals = {}, {}
-    rel_depth = len(inspect.stack()) - depth + 1
-    user_frame = sys._getframe(rel_depth)
+    user_frame = sys._getframe(2)
     user_locals = user_frame.f_locals
     user_globals = user_frame.f_globals
 
     # Look for the required variables
     reqvars = {}
     for var in exprvars:
-        # Get the value.
+        # Get the value
         if var in user_dict:
             val = user_dict[var]
         elif var in user_locals:
@@ -119,11 +113,6 @@ def eval(expression, vm=None, out_flavor=None, user_dict={}, blen=None,
         The length of the block to be evaluated in one go internally.
         The default is a value that has been tested experimentally and
         that offers a good enough peformance / memory usage balance.
-    depth : int
-        The frame depth from which this function is called.  The
-        default is to look up for variables in the outer caller.
-        Make sure that you pass this parameter as
-        ``len(inspect.stack())`` from your own functions.
     kwargs : list of parameters or dictionary
         Any parameter supported by the carray constructor.
 
@@ -151,8 +140,7 @@ def eval(expression, vm=None, out_flavor=None, user_dict={}, blen=None,
         out_flavor = bcolz.defaults.out_flavor
 
     # Get variables and column names participating in expression
-    depth = kwargs.pop('depth', len(inspect.stack()))
-    vars = _getvars(expression, user_dict, depth, vm=vm)
+    vars = _getvars(expression, user_dict, vm=vm)
 
     # Gather info about sizes and lengths
     typesize, vlen = 0, 1