+++ RELEASE_NOTES.rst
@@ -173,9 +173,9 @@ Changes from 1.0.0 to 1.1.0
   object are still kept for backward compatibility, but they are not
   documented anymore and its use is discouraged.
 
-- Added a new `depth` parameter in functions/methods that evaluate
-  expressions so that they can find local and global variables is a more
-  trusted way.
+- Added a new `user_dict` parameter in all ctable methods that evaluate
+  expressions.  For convenience, this dictionary is updated internally
+  with ctable columns, locals and globals from the caller.
 
 
 Changes from 0.12.1 to 1.0.0